+++ NEWS.md
@@ -1,5 +1,9 @@
 # dbplyr (development version)
 
+* `sql_escape_ident()` and `sql_escape_string()` generics from dbplyr are no
+  longer used; these methods should now be supplied for
+  `DBI::dbQuoteIdentifier()` and `DBI::dbQuoteString()` respectively.
+
 * `transmute()` now correctly tracks variables it needs when creating
   subqueries (#313).